data-engineer-airflow-challenge/requirements.txt
apache-airflow[crypto,s3,password,postgres]==1.9.0
psycopg2==2.7.5
requests==2.19.0
python-dotenv==0.15.0
pandas==0.25.3
s3fs==0.5.1